+++ rubato/struct/__init__.py
@@ -1,6 +1,7 @@
 """
 The classes module holds all of the high-dependency modules relating to game objects.
 """
+from .sprite import Sprite
 from .gameobject import *
 from .group import Group
 from .scene import Scene