+++ CHANGELOG.md
@@ -44,6 +44,7 @@
 
 -   Resizing an `Image` now works.
 -   Updated all clone functions to work with new stuff.
+-   Changing the camera zoom now affects all Sprites.
 
 ## [v2.2.0] - June 12, 2022